quarterly data api WIP

+++ hmda-quarterly-data-service/src/main/scala/hmda/quarterly/data/api/dao/AggregatedVolume.scala
@@ -5,5 +5,5 @@ import java.sql.Timestamp
 trait AggregatedVolume {
   val lastUpdated: Timestamp
   val quarter: String
-  val volume: Long
+  val volume: Float
 }
